+++ services/blockchain-indexer/methods/dataService/controllers/legacy.js
@@ -1,0 +1,33 @@
+/*
+ * LiskHQ/lisk-service
+ * Copyright © 2022 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ *
+ */
+const dataService = require('../../../shared/dataService');
+
+const getLegacyAccountInfo = async params => {
+	const legacyAccountInfo = {
+		data: {},
+		meta: {},
+	};
+
+	const response = await dataService.getLegacyAccountInfo(params);
+	if (response.data) legacyAccountInfo.data = response.data;
+	if (response.meta) legacyAccountInfo.meta = response.meta;
+
+	return legacyAccountInfo;
+};
+
+module.exports = {
+	getLegacyAccountInfo,
+};
